add dom to sorageData + factory
