+++ src/factory.js
@@ -1,0 +1,7 @@
+
+const projectFactory = (name) => {
+  const list = [];
+  return { name, list };
+};
+
+export { projectFactory };
